+++ CORBA/CORBA/AccessDB.h
@@ -15,12 +15,12 @@
 #include <mutex>
 
 sql::Connection* getConexion();
-int setUser(sql::Connection* conexion, std::string nombre, std::string correo, std::string pass);
+short setUser(sql::Connection* conexion, std::string nombre, std::string correo, std::string pass);
 bool getUser(sql::Connection* conexion, std::string nombre, std::string pass);
 bool delUser(sql::Connection* conexion, std::string correo);
-int chgPass(sql::Connection* conexion, std::string correo, std::string pass1, std::string pass2);
+short chgPass(sql::Connection* conexion, std::string correo, std::string pass1, std::string pass2);
 P2P::amigos getAmigos(sql::Connection* conexion, std::string correo, P2P::amigos);
-int preAmistad(sql::Connection* conexion, std::string correo1, std::string correo2);
+short preAmistad(sql::Connection* conexion, std::string correo1, std::string correo2);
 bool amistad(sql::Connection* conexion, std::string correo1, std::string correo2);
 P2P::buscar buscar(sql::Connection* conexion, std::string nombre, std::string correo1);
 bool rechazarAmig(sql::Connection* conexion, std::string correo1, std::string correo2);
